fix: 출석 조회 페이지 스타일 분리
